fix(web-app): fallback true, use HOC to add loading
