+++ services/web-app/components/with-loading/index.js
@@ -1,0 +1,8 @@
+/*
+ * Copyright IBM Corp. 2022, 2022
+ *
+ * This source code is licensed under the Apache-2.0 license found in the
+ * LICENSE file in the root directory of this source tree.
+ */
+
+export { default } from './with-loading'
