+++ services/web-app/pages/[...remote-mdx-path].js
@@ -8,6 +8,7 @@
 import { MDXRemote } from 'next-mdx-remote'
 
 import MdxPage from '@/components/mdx-page/mdx-page'
+import WithLoading from '@/components/with-loading'
 import { defaultFilePathPrefix, defaultParams, remotePages } from '@/data/remote-pages'
 import { getRemoteMdxSource } from '@/lib/github'
 import { processMdxSource } from '@/utils/mdx'
@@ -96,8 +97,8 @@ export const getStaticPaths = async () => {
   })
   return {
     paths: allowedPaths,
-    fallback: 'blocking'
+    fallback: true
   }
 }
 
-export default RemoteMdxPage
+export default WithLoading(RemoteMdxPage)
